requirements.txt
torch
torchvision
pytorch-warmup
numpy
pillow
pyyaml
tqdm

